+++ core/deployment/src/main/java/io/quarkus/deployment/Feature.java
@@ -81,6 +81,7 @@ public enum Feature {
     REST_CLIENT_JACKSON,
     REST_CLIENT_JAXB,
     REST_CLIENT_JSONB,
+    REST_CLIENT_MUTINY,
     SCALA,
     SCHEDULER,
     SECURITY,
