+++ extensions/pom.xml
@@ -50,12 +50,14 @@
         <module>resteasy-jsonb</module>
         <module>resteasy-jackson</module>
         <module>resteasy-jaxb</module>
+        <module>resteasy-mutiny-common</module>
         <module>resteasy-mutiny</module>
         <module>resteasy-qute</module>
         <module>rest-client</module>
         <module>rest-client-jackson</module>
         <module>rest-client-jsonb</module>
         <module>rest-client-jaxb</module>
+        <module>rest-client-mutiny</module>
         <module>smallrye-openapi-common</module>
         <module>smallrye-openapi</module>
         <module>swagger-ui</module>
