+++ extensions/rest-client-mutiny/deployment/src/main/java/io/quarkus/restclient/mutiny/deployment/RestClientMutinyProcessor.java
@@ -1,14 +1,14 @@
-package io.quarkus.resteasy.mutiny.deployment;
+package io.quarkus.restclient.mutiny.deployment;
 
 import io.quarkus.deployment.Feature;
 import io.quarkus.deployment.annotations.BuildStep;
 import io.quarkus.deployment.builditem.FeatureBuildItem;
 
-public class ResteasyMutinyProcessor {
+public class RestClientMutinyProcessor {
 
     @BuildStep
     FeatureBuildItem feature() {
-        return new FeatureBuildItem(Feature.RESTEASY_MUTINY);
+        return new FeatureBuildItem(Feature.REST_CLIENT_MUTINY);
     }
 
 }
