+++ extensions/rest-client-mutiny/deployment/src/test/java/io/quarkus/restclient/mutiny/deployment/MutinyRestClientTest.java
@@ -1,16 +1,13 @@
-package io.quarkus.restclient.ft;
+package io.quarkus.restclient.mutiny.deployment;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.net.URL;
+import java.util.concurrent.ExecutionException;
 
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
-import javax.ws.rs.WebApplicationException;
 
-import org.eclipse.microprofile.faulttolerance.ExecutionContext;
-import org.eclipse.microprofile.faulttolerance.Fallback;
-import org.eclipse.microprofile.faulttolerance.FallbackHandler;
 import org.eclipse.microprofile.rest.client.RestClientBuilder;
 import org.eclipse.microprofile.rest.client.inject.RegisterRestClient;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
@@ -20,30 +17,30 @@ import org.junit.jupiter.api.extension.RegisterExtension;
 
 import io.quarkus.test.QuarkusUnitTest;
 import io.quarkus.test.common.http.TestHTTPResource;
+import io.smallrye.mutiny.Uni;
 
-public class RestClientFallbackTest {
+public class MutinyRestClientTest {
 
     @RegisterExtension
     static final QuarkusUnitTest config = new QuarkusUnitTest()
             .setArchiveProducer(() -> ShrinkWrap.create(JavaArchive.class)
-                    .addClasses(RestClientFallbackTest.class, Client.class, MyFallback.class));
+                    .addClasses(MutinyRestClientTest.class, Client.class, TestEndpoint.class));
 
     @TestHTTPResource
     URL url;
 
     @Test
-    public void testFallbackWasUsed() {
+    public void testUni() throws InterruptedException, ExecutionException {
         Client client = RestClientBuilder.newBuilder().baseUrl(url).build(Client.class);
-        assertEquals("pong", client.ping());
+        assertEquals("OK", client.ping().await().indefinitely());
     }
 
     @RegisterRestClient
     public interface Client {
 
-        @Fallback(MyFallback.class)
         @GET
         @Path("/test")
-        String ping();
+        Uni<String> ping();
 
     }
 
@@ -52,16 +49,7 @@ public class RestClientFallbackTest {
 
         @GET
         public String get() {
-            throw new WebApplicationException(404);
-        }
-
-    }
-
-    public static class MyFallback implements FallbackHandler<String> {
-
-        @Override
-        public String handle(ExecutionContext context) {
-            return "pong";
+            return "OK";
         }
 
     }
