+++ extensions/rest-client-mutiny/runtime/src/main/resources/META-INF/quarkus-extension.yaml
@@ -1,0 +1,13 @@
+---
+name: "Mutiny support for REST Client"
+metadata:
+  keywords:
+  - "rest-client-mutiny"
+  - "rest-client"
+  - "web-client"
+  - "microprofile-rest-client"
+  - "mutiny"
+  categories:
+  - "web"
+  - "reactive"
+  status: "preview"
