--- extensions/resteasy-mutiny/runtime/src/main/java/io/quarkus/resteasy/mutiny/runtime/MultiProvider.java
@@ -1,13 +1,0 @@
-package io.quarkus.resteasy.mutiny.runtime;
-
-import org.jboss.resteasy.spi.AsyncStreamProvider;
-import org.reactivestreams.Publisher;
-
-import io.smallrye.mutiny.Multi;
-
-public class MultiProvider implements AsyncStreamProvider<Multi<?>> {
-    @Override
-    public Publisher<?> toAsyncStream(Multi<?> multi) {
-        return multi;
-    }
-}
