--- extensions/resteasy-mutiny/runtime/src/main/java/io/quarkus/resteasy/mutiny/runtime/MultiRxInvoker.java
@@ -1,14 +1,0 @@
-package io.quarkus.resteasy.mutiny.runtime;
-
-import javax.ws.rs.client.RxInvoker;
-
-import io.smallrye.mutiny.Multi;
-import io.smallrye.mutiny.subscription.BackPressureStrategy;
-
-public interface MultiRxInvoker extends RxInvoker<Multi<?>> {
-
-    BackPressureStrategy getBackPressureStrategy();
-
-    void setBackPressureStrategy(BackPressureStrategy strategy);
-
-}
